Fixed player respawn after battle
Needed to use this: https://forum.unity.com/threads/how-to-use-scenemanager-onsceneloaded.399221/ because player was getting transform before the scene change. Now, on successful scene load the transform is changed to the orignal position. Also, needed to hardcode camera bounds into first room and then compare on start in the camera to make sure the camera bounds were the same as before the scene change. Unity being unity